Draft PlanningService

+++ src/main/java/com/asialocalguide/gateway/core/dto/ActivityPlanningRequestDTO.java
@@ -4,7 +4,4 @@ import java.time.Instant;
 import java.util.List;
 
 public record ActivityPlanningRequestDTO(
-    Instant startDateISO,
-    Instant endDateISO,
-    Integer destinationId,
-    List<Integer> activityTagIds) {}
+    Instant startDateISO, Instant endDateISO, Long destinationId, List<Long> activityTagIds) {}
